ACR-690 Crime Version Update (#130)
* Removing query usage for GET crime-version endpoint in favour of entity relationships

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringcrimematchingapi/dto/DeviceWearerPositionResponse.kt
@@ -2,10 +2,10 @@ package uk.gov.justice.digital.hmpps.electronicmonitoringcrimematchingapi.dto
 
 data class DeviceWearerPositionResponse(
   val capturedDateTime: String,
-  val direction: Int,
+  val direction: Long,
   val latitude: Double,
   val longitude: Double,
-  val precision: Int,
+  val precision: Long,
   val sequenceLabel: String,
-  val speed: Int,
+  val speed: Long,
 )
